requirements.txt
langchain
openai
streamlit
python-dotenv
youtube-transcript-api
faiss-cpu